bot/requirements.txt
aiogram

# settings
python-dotenv
pydantic-settings

faststream[rabbit]
